Going for fastest zombie
So added a little so they would shoot at the fastest zombie. Changed a couple variables. the fovAngle I changed to see 360 degrees because there were rocks right next to them that they were not picking up simply because it wasn't in their vision. Created a collection range for the rocks so that they will go after rocks that are farther away, if needed this can be taken out once we get the walls done. Also changed the vision of the hunters to 300. 

The walls I am still trying to figure out....so far haven't found the right way to get it to push at an angle so hunters go in a circle.